+++ services/sms/src/main/java/com/huaweicloud/sdk/sms/v3/model/PhysicalVolumes.java
@@ -51,13 +51,13 @@ public class PhysicalVolumes  {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="size")
     
-    private Integer size;
+    private Long size;
 
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="used_size")
     
-    private Integer usedSize;
+    private Long usedSize;
 
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -167,7 +167,7 @@ public class PhysicalVolumes  {
         this.name = name;
     }
 
-    public PhysicalVolumes withSize(Integer size) {
+    public PhysicalVolumes withSize(Long size) {
         this.size = size;
         return this;
     }
@@ -180,15 +180,15 @@ public class PhysicalVolumes  {
      * minimum: 0
      * @return size
      */
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(Long size) {
         this.size = size;
     }
 
-    public PhysicalVolumes withUsedSize(Integer usedSize) {
+    public PhysicalVolumes withUsedSize(Long usedSize) {
         this.usedSize = usedSize;
         return this;
     }
@@ -201,11 +201,11 @@ public class PhysicalVolumes  {
      * minimum: 0
      * @return usedSize
      */
-    public Integer getUsedSize() {
+    public Long getUsedSize() {
         return usedSize;
     }
 
-    public void setUsedSize(Integer usedSize) {
+    public void setUsedSize(Long usedSize) {
         this.usedSize = usedSize;
     }
