+++ services/sms/src/main/java/com/huaweicloud/sdk/sms/v3/model/TargetDisk.java
@@ -136,13 +136,13 @@ public class TargetDisk  {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="size")
     
-    private Integer size;
+    private Long size;
 
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value="used_size")
     
-    private Integer usedSize;
+    private Long usedSize;
 
     public TargetDisk withDeviceUse(DeviceUseEnum deviceUse) {
         this.deviceUse = deviceUse;
@@ -238,7 +238,7 @@ public class TargetDisk  {
         this.physicalVolumes = physicalVolumes;
     }
 
-    public TargetDisk withSize(Integer size) {
+    public TargetDisk withSize(Long size) {
         this.size = size;
         return this;
     }
@@ -250,15 +250,15 @@ public class TargetDisk  {
      * 大小
      * @return size
      */
-    public Integer getSize() {
+    public Long getSize() {
         return size;
     }
 
-    public void setSize(Integer size) {
+    public void setSize(Long size) {
         this.size = size;
     }
 
-    public TargetDisk withUsedSize(Integer usedSize) {
+    public TargetDisk withUsedSize(Long usedSize) {
         this.usedSize = usedSize;
         return this;
     }
@@ -270,11 +270,11 @@ public class TargetDisk  {
      * 已使用大小
      * @return usedSize
      */
-    public Integer getUsedSize() {
+    public Long getUsedSize() {
         return usedSize;
     }
 
-    public void setUsedSize(Integer usedSize) {
+    public void setUsedSize(Long usedSize) {
         this.usedSize = usedSize;
     }
     @Override
